pyproject.toml
[tool.poetry]
name = "databaseconverter"
version = "0.1.0"
description = "Conversor de banco de dados"
authors = ["Anderson Rezende <andersonrezende17@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
mysql-connector-python = "^8.2.0"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
blue = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
